[Bronze V] Title: A/B, Time: 0 ms, Memory: 1112 KB -BaekjoonHub

+++ 백준/Bronze/1008. A／B/A／B.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    float a, b;
-    scanf("%f %f",&a,&b);
-    printf("%.9f",(double)a/b);
+    double a, b;
+    scanf("%lf %lf",&a,&b);
+    printf("%.9lf",a/b);
     return 0;
 }
